requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0
groq>=0.11
mcp[cli]>=1.8
office-powerpoint-mcp-server>=2.0.7
